--- config/etc/addrtab/v7xx/kerberos_mib_v2/ipbus_freq_ctr.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-
-<node description="Frequency counter" fwinfo="endpoint;width=1">
-	<node id="ctrl" address="0x0">
-		<node id="chan_sel" mask="0xf"/>
-		<node id="en_crap_mode" mask="0x10"/>
-	</node>
-	<node id="freq" address="0x1">
-		<node id="count" mask="0xffffff"/>
-		<node id="valid" mask="0x1000000"/>
-	</node>
-</node>
