--- config/etc/addrtab/v7xx/kerberos_mib_v2/opencores_i2c.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-
-<node description="I2C master controller" fwinfo="endpoint;width=3">
-	<node id="ps_lo" address="0x0" description="Prescale low byte"/>
-	<node id="ps_hi" address="0x1" description="Prescale low byte"/>
-	<node id="ctrl" address="0x2" description="Control"/>
-	<node id="data" address="0x3" description="Data"/>
-	<node id="cmd_stat" address="0x4" description="Command / status"/>
-</node>
