--- config/etc/addrtab/v7xx/kerberos_mib_v2/pdts_ep_mon.xml
@@ -1,26 +1,0 @@
-<node id="ep_core" description="PDTS endpoint core wrapper" fwinfo="endpoint" class="EndpointNode">
-		<node id="csr" address="0x0" fwinfo="endpoint;width=1">
-			<node id="ctrl" address="0x0">
-				<node id="ep_en" mask="0x1"/>
-				<node id="ctr_rst" mask="0x2"/>
-				<node id="addr" mask="0xffff0000"/>
-			</node>
-			<node id="stat" address="0x1">
-				<node id="ep_stat" mask="0x0f"/>
-				<node id="ep_rdy" mask="0x10"/>
-				<node id="ep_txen" mask="0x20"/>
-				<node id="ctrs_rdy" mask="0x40"/>
-			</node>
-		</node>
-		<node id="tstamp" address="0x2" mode="block" size="0x2" fwinfo="endpoint;width=1"/>
-		<node id="cmd_ctrs" address="0x4" fwinfo="endpoint;width=1">
-			<node id="addr" address="0x0"/>
-			<node id="data" address="0x1" size="0x100" mode="port"/>
-		</node>
-		<node id="cmd_log" address="0x8" fwinfo="endpoint;width=2">
-			<node id="tstamp_l" address="0x0"/>
-			<node id="tstamp_h" address="0x1"/>
-			<node id="cmd" address="0x2"/>
-		</node>
-		<node id="dcmd_ctrs" address="0x10" mode="block" size="0x8" fwinfo="endpoint;width=3"/>
-</node>
