--- config/etc/addrtab/v7xx/kerberos_mib_v2/pdts_master.xml
@@ -1,24 +1,0 @@
-<node id="pdts_master" description="PDTS master block" fwinfo="endpoint" class="MasterNode">
-		<node id="global" address="0x0" module="file://pdts_global.xml"/>
-		<node id="acmd_buf" address="0x4" fwinfo="endpoint;width=2">
-			<node id="stat" address="0x0" permission="r">
-				<node id="ready" mask="0x1"/>
-				<node id="timeout" mask="0x2"/>
-			</node>
-			<node id="txbuf" address="0x2" mode="port" size="0x20" permission="w"/>
-			<node id="rxbuf" address="0x3" mode="port" size="0x20" permission="r"/>			
-		</node>
-		<node id="cmd_log" address="0x8" fwinfo="endpoint;width=2">
-			<node id="tstamp_l" address="0x0"/>
-			<node id="tstamp_h" address="0x1"/>
-			<node id="cmd" address="0x2"/>
-		</node>
-		<node id="cmd_ctrs" address="0xc" fwinfo="endpoint;width=1">
-			<node id="addr" address="0x0"/>
-			<node id="data" address="0x1" size="0x100" mode="port"/>
-		</node>
-		<node id="echo_mon" address="0xe" module="file://pdts_echo_mon.xml"/>
-		<node id="tstamp" address="0x10" module="file://pdts_ts_source.xml"/>
-		<node id="scmd_gen" address="0x40" module="file://pdts_scmd_gen.xml"/>
-</node>
-
