--- config/etc/addrtab/v7xx/kerberos_mib_v2/pdts_mib_v2_io.xml
@@ -1,36 +1,0 @@
-<node id="io" description="MIB v2 IO" fwinfo="endpoint" class="MIBV2IONode">
-		<node id="csr" address="0x0" description="ctrl/stat register" fwinfo="endpoint;width=1">
-				<node id="ctrl" address="0x0">
-					<node id="soft_rst" mask="0x1"/>
-					<node id="nuke" mask="0x2"/>
-					<node id="rst" mask="0x4"/>
-					<node id="pll_in_sel" mask="0x18"/>
-					<node id="bp_tx_edge" mask="0x20"/>
-					<node id="usfp_tx_edge" mask="0x40"/>
-				</node>
-				<node id="stat" address="0x1">
-					<node id="sfp0_flt" mask="0x1"/>
-					<node id="sfp1_flt" mask="0x2"/>        	
-					<node id="sfp2_flt" mask="0x4"/>
-					<node id="sfp0_los" mask="0x8"/>
-					<node id="sfp1_los" mask="0x10"/>        	
-					<node id="sfp2_los" mask="0x20"/>
-					<node id="pll_ok" mask="0x40"/>
-					<node id="pll_sticky" mask="0x80"/>
-					<node id="mmcm_ok" mask="0x100"/>
-					<node id="mmcm_sticky" mask="0x200"/>
-				</node>
-		</node>
-		<node id="config" address="0x2" description="board/firmware config" fwinfo="endpoint;width=1">
-		    <node id="no_cdr" mask="0xff000000"/>
-		    <node id="board_type" mask="0xff0000"/>
-		    <node id="carrier_type" mask="0xff00"/>
-		    <node id="design_type" mask="0xff"/>
-		    <node id="clock_frequency" address="0x1" mask="0xffffffff"/>
-		</node>
-		<node id="i2c" address="0x08" module="file://opencores_i2c.xml" class="I2CMasterNode" parameters="UID_PROM=0x50;PLL=0x68"/>
-		<node id="sfp0_i2c" address="0x10" module="file://opencores_i2c.xml" class="I2CMasterNode" parameters="SFP_EEProm=0x50;SFP_Diag=0x51"/>
-		<node id="sfp1_i2c" address="0x20" module="file://opencores_i2c.xml" class="I2CMasterNode" parameters="SFP_EEProm=0x50;SFP_Diag=0x51"/>
-		<node id="sfp2_i2c" address="0x40" module="file://opencores_i2c.xml" class="I2CMasterNode" parameters="SFP_EEProm=0x50;SFP_Diag=0x51"/>
-		<node id="freq" address="0x50" module="file://ipbus_freq_ctr.xml" class="FrequencyCounterNode"/>
-</node>
