--- config/etc/addrtab/v7xx/kerberos_mib_v2/pdts_timestamp_comparator.xml
@@ -1,16 +1,0 @@
-<node id="pdts_timestamp_comparator" description="Comparator of 64b timestamps" fwinfo="endpoint">
-		<node id="csr" address="0x0" fwinfo="endpoint;width=5">
-			<node id="ctrl" address="0x0">
-				<node id="rst" mask="0x1"/>
-				<node id="go" mask="0x2"/>
-			</node>
-			<node id="stat" address="0x1">
-				<node id="err_sticky" mask="0x1"/>
-				<node id="valid" mask="0x2"/>
-			</node>
-			<node id="ts_0_lo" address="0x2"/>
-			<node id="ts_0_hi" address="0x3"/>
-			<node id="ts_1_lo" address="0x4"/>
-			<node id="ts_1_hi" address="0x5"/>
-		</node>
-</node>
