--- config/etc/addrtab/v7xx/kerberos_mib_v2/pdts_timing_source_mux.xml
@@ -1,7 +1,0 @@
-<node id="pdts_timing_source_mux" description="Switch between upstream timing sources" fwinfo="endpoint">
-		<node id="csr" address="0x0" fwinfo="endpoint;width=0">
-			<node id="ctrl" address="0x0">
-				<node id="src" mask="0x3"/>
-			</node>
-		</node>
-</node>
