--- config/etc/addrtab/v7xx/kerberos_mib_v2/pdts_ts_source.xml
@@ -1,21 +1,0 @@
-<!-- https://dunescience.slack.com/archives/G01CH2K070X/p1684404277486259 -->
-
-<node id="pdts_ts_source" description="Timestamp generator" fwinfo="endpoint" class="TimestampGeneratorNode">
-  <node id="csr" address="0x0" fwinfo="endpoint;width=3">
-    <node id="ctrl" address="0x0">
-      <node id="rst" mask="0x1"/>
-      <node id="load" mask="0x2"/>
-      <node id="tstamp_source_sel" mask="0xc"/>
-    </node>
-    <node id="tstamp_sw_init_l" address="0x1"/>
-    <node id="tstamp_sw_init_h" address="0x2"/>
-    <node id="stat" address="0x4">
-      <node id="tstamp_loaded" mask="0x1"/>
-      <node id="tstamp_error" mask="0x2"/>
-    </node>
-    <node id="tstamp_start_l" address="0x5"/>
-    <node id="tstamp_start_h" address="0x6"/>
-  </node>
-  <node id="ctr" mode="block" size="0x2" address="0x8" fwinfo="endpoint;width=2"/>
-</node>
-
